tools/scripts/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
numpy>=1.24.0
git+https://github.com/openai/CLIP.git
